+++ .github/workflows/test.yml
@@ -36,6 +36,9 @@ jobs:
         # E203/W503/W504 = this is where black and flake8 conflict, see https://black.readthedocs.io/en/stable/faq.html#why-are-flake8-s-e203-and-w503-violated
         flake8 --ignore=E501,E203,W503 --extend-select=W504
 
+    - name: Check types
+      run: mypy src tests
+
     - name: Run tests
       run: |
         coverage run -m pytest test_app.py test_referrers.py test_utils.py
