+++ .gitignore
@@ -1,3 +1,4 @@
+*.egg-info
 *.pyc
 *.sqlite
 .coverage
