+++ src/analytics/__init__.py
@@ -1,0 +1,1 @@
+from .app import app
