+++ src/analytics/static/a.gif
