+++ src/analytics/static/dashboard.js
