+++ src/analytics/static/style.css
