+++ src/analytics/templates/base.html
