+++ src/analytics/templates/dashboard.html
