+++ src/analytics/templates/index.html
